requirements.txt
Django>=3.2.16
djangorestframework==3.13.1
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
djangorestframework-simplejwt>=5.2.2,<5.3.0
django-filter==2.4.0
django-ses==2.6.1
django-enumchoicefield==3.0.1